Price is `int` now

+++ book_RL_Stanford/001/script.py
@@ -34,6 +34,6 @@ print(
                 100,
             )
         ),
-        float,
+        int,
     ),
 )
